Add FE for importing recipe images and clean up a little.
Also remove a missed console.log from the edit test.
