+++ packages/component-library/components/Form/inputs/File/index.tsx
@@ -1,24 +1,27 @@
-import { ChangeEventHandler } from "react";
+import { ChangeEventHandler, forwardRef, ForwardedRef } from "react";
 import clsx from "clsx";
 import { Errors, FieldWrapper, baseInputStyle } from "../..";
 
-export function FileInput({
-  name,
-  id = name,
-  defaultValue,
-  onChange,
-  label,
-  placeholder,
-  errors,
-}: {
-  name: string;
-  id?: string;
-  label: string;
-  defaultValue?: string;
-  onChange?: ChangeEventHandler<HTMLInputElement>;
-  placeholder?: string;
-  errors?: string[];
-}) {
+function FileInputComponent(
+  {
+    name,
+    id = name,
+    defaultValue,
+    onChange,
+    label,
+    placeholder,
+    errors,
+  }: {
+    name: string;
+    id?: string;
+    label: string;
+    defaultValue?: string;
+    onChange?: ChangeEventHandler<HTMLInputElement>;
+    placeholder?: string;
+    errors?: string[];
+  },
+  ref?: ForwardedRef<HTMLInputElement>,
+) {
   return (
     <FieldWrapper label={label} id={id}>
       <Errors errors={errors} />
@@ -30,8 +33,11 @@ export function FileInput({
           defaultValue={defaultValue}
           onChange={onChange}
           placeholder={placeholder}
+          ref={ref}
         />
       </div>
     </FieldWrapper>
   );
 }
+
+export const FileInput = forwardRef(FileInputComponent);
